1. adding etx_device cdev in module init 2. fix missing cdev after adding in probe issue (see devcie_create, devno)
I first add etx_device cdev to test wheter it is invalid to add cdev in probe. this idea comes from those exmple I saw in the past. They all add the cdev in module_init.
=> this approch is able to add cdev.

After trying, the cdev is finally able to be added in probe. the root cause is that I accidently give the major number to the device_create(). it is the full device number that device_create() needs, not the major number. device_create() will get the major part itself by right shifting.